requirements.txt
certifi==2019.11.28
chardet==3.0.4
cx-Freeze==6.0
idna==2.8
PyJWT==1.7.1
requests==2.22.0
urllib3==1.25.7
